docs: autogenerate documentation pour Python modules
