--- docs/developers/modules/cogip/config.md
@@ -1,1 +1,0 @@
-::: cogip.config
