--- docs/developers/modules/cogip/entities/asset.md
@@ -1,1 +1,0 @@
-::: cogip.entities.asset
