--- docs/developers/modules/cogip/entities/dynobstacle.md
@@ -1,1 +1,0 @@
-::: cogip.entities.dynobstacle
