--- docs/developers/modules/cogip/entities/impact.md
@@ -1,1 +1,0 @@
-::: cogip.entities.impact
