--- docs/developers/modules/cogip/entities/line.md
@@ -1,1 +1,0 @@
-::: cogip.entities.line
