--- docs/developers/modules/cogip/entities/obstacle.md
@@ -1,1 +1,0 @@
-::: cogip.entities.obstacle
