--- docs/developers/modules/cogip/entities/path.md
@@ -1,1 +1,0 @@
-::: cogip.entities.path
