--- docs/developers/modules/cogip/entities/robot.md
@@ -1,1 +1,0 @@
-::: cogip.entities.robot
