--- docs/developers/modules/cogip/entities/sensor.md
@@ -1,1 +1,0 @@
-::: cogip.entities.sensor
