--- docs/developers/modules/cogip/entities/table.md
@@ -1,1 +1,0 @@
-::: cogip.entities.table
