--- docs/developers/modules/cogip/models/models.md
@@ -1,1 +1,0 @@
-::: cogip.models.models
