--- docs/developers/modules/cogip/tools/copilot/app.md
@@ -1,1 +1,0 @@
-::: cogip.tools.copilot.app
