--- docs/developers/modules/cogip/tools/copilot/copilot.md
@@ -1,1 +1,0 @@
-::: cogip.tools.copilot.copilot
