--- docs/developers/modules/cogip/tools/copilot/main.md
@@ -1,1 +1,0 @@
-::: cogip.tools.copilot.__main__
