--- docs/developers/modules/cogip/tools/copilot/message_types.md
@@ -1,1 +1,0 @@
-::: cogip.tools.copilot.message_types
