--- docs/developers/modules/cogip/tools/copilot/queues.md
@@ -1,1 +1,0 @@
-::: cogip.tools.copilot.queues
