--- docs/developers/modules/cogip/tools/lidarpf/dataproxy.md
@@ -1,1 +1,0 @@
-::: cogip.tools.lidarpf.dataproxy
