--- docs/developers/modules/cogip/tools/lidarpf/lidarserial.md
@@ -1,1 +1,0 @@
-::: cogip.tools.lidarpf.lidarserial
