--- docs/developers/modules/cogip/tools/lidarpf/main.md
@@ -1,1 +1,0 @@
-::: cogip.tools.lidarpf.__main__
