--- docs/developers/modules/cogip/tools/lidarpf/mainwindow.md
@@ -1,1 +1,0 @@
-::: cogip.tools.lidarpf.mainwindow
