--- docs/developers/modules/cogip/tools/lidarusb/dataproxy.md
@@ -1,1 +1,0 @@
-::: cogip.tools.lidarusb.dataproxy
