--- docs/developers/modules/cogip/tools/lidarusb/lidarserial.md
@@ -1,1 +1,0 @@
-::: cogip.tools.lidarusb.lidarserial
