--- docs/developers/modules/cogip/tools/lidarusb/main.md
@@ -1,1 +1,0 @@
-::: cogip.tools.lidarusb.__main__
