--- docs/developers/modules/cogip/tools/lidarusb/mainwindow.md
@@ -1,1 +1,0 @@
-::: cogip.tools.lidarusb.mainwindow
