--- docs/developers/modules/cogip/tools/monitor/main.md
@@ -1,1 +1,0 @@
-::: cogip.tools.monitor.__main__
