--- docs/developers/modules/cogip/tools/monitor/mainwindow.md
@@ -1,1 +1,0 @@
-::: cogip.tools.monitor.mainwindow
