--- docs/developers/modules/cogip/tools/monitor/socketiocontroller.md
@@ -1,1 +1,0 @@
-::: cogip.tools.monitor.socketiocontroller
