--- docs/developers/modules/cogip/tools/replay/main.md
@@ -1,1 +1,0 @@
-::: cogip.tools.replay.__main__
