--- docs/developers/modules/cogip/tools/replay/mainwindow.md
@@ -1,1 +1,0 @@
-::: cogip.tools.replay.mainwindow
