--- docs/developers/modules/cogip/utils/lidartablemodel.md
@@ -1,1 +1,0 @@
-::: cogip.utils.lidartablemodel
