--- docs/developers/modules/cogip/widgets/chartsview.md
@@ -1,1 +1,0 @@
-::: cogip.widgets.chartsview
