--- docs/developers/modules/cogip/widgets/gameview.md
@@ -1,1 +1,0 @@
-::: cogip.widgets.gameview
