--- docs/developers/modules/cogip/widgets/lidarview.md
@@ -1,1 +1,0 @@
-::: cogip.widgets.lidarview
