+++ docs/requirements.txt
@@ -1,5 +1,7 @@
 MarkupSafe==2.0.1
 mkdocs==1.2.3
+mkdocs-gen-files==0.3.3
+mkdocs-literate-nav==0.4.1
 mkdocs-material==8.1.3
 mkdocstrings==0.16.2
 pymdown-extensions==9.1
